changing cloudinit solution name to be virtual machine.
