+++ src/App.svelte
@@ -1,7 +1,7 @@
 <script lang="ts">
   import Base from "./elements/base/Base.wc.svelte";
   import ProfileManager from "./elements/profiles/Profiles.wc.svelte";
-  import VM from "./elements/vm/Vm.wc.svelte";
+  import VM from "./elements/microVM/Vm.wc.svelte";
 </script>
 
 <Base />
