+++ src/components/LogsInfo.svelte
@@ -16,12 +16,12 @@
         return `Logs of ${s("Removing Worker")} with name ${s(name)}`;
 
       case "VM":
+      case "MicroVM":
       case "CapRover":
       case "Funkwhale":
       case "Kubernetes":
       case "Peertube":
       case "Owncloud":
-      case "CloudInit":
       case "GatewayName":
         return `Logs of Deployment of type ${s(type)} with name ${s(name)}`;
