--- src/elements/cloudInit/index.ts
@@ -1,1 +1,0 @@
-import "./CloudInit.wc.svelte";
