+++ src/elements/microVM/index.ts
@@ -1,0 +1,1 @@
+import "./Vm.wc.svelte";
