+++ src/types/fullVM.ts
